videoapi/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=videoapi/requirements.txt --resolver=backtracking req/pyproject.toml
#
aiohttp==3.8.4
    # via VideoApi (req/pyproject.toml)
aiosignal==1.3.1
    # via aiohttp
anyio==3.6.2
    # via
    #   httpcore
    #   starlette
async-timeout==4.0.2
    # via aiohttp
asyncio==3.4.3
    # via VideoApi (req/pyproject.toml)
attrs==23.1.0
    # via aiohttp
azure-core==1.26.3
    # via azure-storage-blob
azure-functions==1.14.0
    # via VideoApi (req/pyproject.toml)
azure-storage-blob==12.16.0
    # via VideoApi (req/pyproject.toml)
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.0.1
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via uvicorn
colorama==0.4.6
    # via click
cryptography==39.0.1
    # via azure-storage-blob
fastapi==0.95.1
    # via VideoApi (req/pyproject.toml)
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.16.3
    # via httpx
httpx==0.24.0
    # via VideoApi (req/pyproject.toml)
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
isodate==0.6.1
    # via azure-storage-blob
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
pycparser==2.21
    # via cffi
pydantic==1.10.5
    # via fastapi
python-multipart==0.0.6
    # via VideoApi (req/pyproject.toml)
requests==2.28.2
    # via azure-core
six==1.16.0
    # via
    #   azure-core
    #   isodate
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.26.1
    # via fastapi
typing-extensions==4.5.0
    # via
    #   azure-core
    #   azure-storage-blob
    #   pydantic
urllib3==1.26.14
    # via requests
uvicorn==0.22.0
    # via VideoApi (req/pyproject.toml)
yarl==1.9.2
    # via aiohttp
